Some commenting , added HubQuit custom message

+++ src/com/phinvader/libjdcpp/DCConstants.java
@@ -10,8 +10,8 @@ public class DCConstants {
 	public static String version = "libjdcpp v1.0";
 
 	/**
-	 * This is the number of bytes per chunk sent while transfering files over
-	 * the TCP connection
+	 * This is the number of bytes per chunk sent while transferring files over
+	 * the TCP connection. (As specified by the DC protocol)
 	 */
 	public static long data_chunk_size = 40906;
 	/**
@@ -19,5 +19,5 @@ public class DCConstants {
 	 * MessageHandler Instance. If more messages are arrive then the TCP Socket
 	 * will block and wait till the message Queue is freed
 	 */
-	public static long max_message_queue_size = 1024;
+	public static int max_message_queue_size = 1024;
 }
